3 down. 3 to go.
I spent a little more time on the “Interactive” portion of the array as
I re-taught myself on how to prompt a number of times to make a custom
array.

So far, this is the funnest assignment we’ve had in SDI. I love
especially am loving getting creative with the “base code” portion to
make it so you can interact with it and use your own variable data if
you’d like. I know that’s not graded, but I’m confident that I’ll have
the other functions done with plenty of time.